feat/chore: bsearch in getByName/getById and fix input parameter type of getById to make it same as whaleId

+++ src/main/java/ca/uvic/seng330/ex3/ObservationRepository.java
@@ -55,7 +55,7 @@ public class ObservationRepository implements Repository<Observation>{
      * @return a observation on whale with id <code>id</code>
      */
     @Override
-    public Observation getById(int id) {
+    public Observation getById(long id) {
         System.out.println("Getting Observations by Species");
         return null;
     }
